:art: 将部分重复代码抽象出来
  解决了部分亢余代码。

:art: 将部分重复代码抽象出来

  解决了部分亢余代码。

:art: 将部分重复代码抽象出来

  解决了部分亢余代码。

:art: 将部分重复代码抽象出来

  解决了部分亢余代码。

:art: 将部分重复代码抽象出来

  解决了部分亢余代码。
