+++ module_todo/src/main/java/com/mredrock/cyxbs/todo/ui/fragment/TodoLifeFragment.kt
@@ -6,7 +6,6 @@ import android.os.Build
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -15,7 +14,6 @@ import android.widget.FrameLayout
 import androidx.annotation.RequiresApi
 import androidx.appcompat.widget.LinearLayoutCompat
 import androidx.fragment.app.activityViewModels
-import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
@@ -32,12 +30,8 @@ import com.mredrock.cyxbs.todo.model.bean.TodoListPushWrapper
 import com.mredrock.cyxbs.todo.model.bean.TodoPinData
 import com.mredrock.cyxbs.todo.ui.activity.TodoDetailActivity
 import com.mredrock.cyxbs.todo.ui.dialog.DeleteTodoDialog
+import com.mredrock.cyxbs.todo.util.TodoHelper.updateTodoItem
 import com.mredrock.cyxbs.todo.viewmodel.TodoViewModel
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-import java.time.LocalDateTime
-import java.time.format.DateTimeFormatter
 
 /**
  * description ：清单下面四个页面之一
@@ -58,9 +52,6 @@ class TodoLifeFragment : BaseFragment(), TodoAllAdapter.OnItemClickListener {
     private val handler = Handler(Looper.getMainLooper())
     private var pendingUpdateTask: Runnable? = null
 
-    @RequiresApi(Build.VERSION_CODES.O)
-    private val dateTimeFormatter: DateTimeFormatter =
-        DateTimeFormatter.ofPattern("yyyy年M月d日HH:mm")
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -116,9 +107,6 @@ class TodoLifeFragment : BaseFragment(), TodoAllAdapter.OnItemClickListener {
                         )
 
                     )
-                    for (item in todoAllAdapter.selectItems) {
-                        Log.d("SwipeDeleteRecyclerView", "deletePinning item: ${item.todoId}")
-                    }
                     todoAllAdapter.selectItems.clear()
                     dismiss()
                 }.setNegativeClick {
@@ -128,7 +116,6 @@ class TodoLifeFragment : BaseFragment(), TodoAllAdapter.OnItemClickListener {
         }
         acTopButton.setOnClickListener {
             val syncTime = appContext.getSp("todo").getLong("TODO_LAST_SYNC_TIME", 0L)
-            Log.d("SwipeDeleteRecyclerView", "Pinning item: ${todoAllAdapter.selectItems}")
 
             for (item in todoAllAdapter.selectItems) {
                 mViewModel.pinTodo(TodoPinData(1, 1, syncTime.toInt(), item.todoId.toInt()))
@@ -146,7 +133,6 @@ class TodoLifeFragment : BaseFragment(), TodoAllAdapter.OnItemClickListener {
 
     private fun ifClick() {
         mViewModel.isEnabled.observe(viewLifecycleOwner) {
-            Log.d("TodoAllFragment", "isEnabled changed: $it")
             if (it) {
                 showBatchManagementLayout()
             } else {
@@ -169,7 +155,6 @@ class TodoLifeFragment : BaseFragment(), TodoAllAdapter.OnItemClickListener {
     }
 
     private fun checkIfEmpty() {
-        Log.d("TodoAllFragment", "Checking if empty. Item count: ${todoAllAdapter.itemCount}")
         if (todoAllAdapter.itemCount == 0) {
             mRecyclerView.visibility = View.GONE
             emptyview.visibility = View.VISIBLE
@@ -219,10 +204,6 @@ class TodoLifeFragment : BaseFragment(), TodoAllAdapter.OnItemClickListener {
                 }.show()
 
 
-        } else {
-            // 如果索引无效，记录错误日志
-            Log.e("TodoAllFragment", "Invalid index: $position, Size: ${currentList.size}")
-
         }
     }
 
@@ -259,316 +240,21 @@ class TodoLifeFragment : BaseFragment(), TodoAllAdapter.OnItemClickListener {
         }
     }
 
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun updateTodoItem(todoItem: Todo) {
-        Log.d("day", "updateTodoItem: ${todoItem.endTime}")
-        if (todoItem.endTime == todoItem.remindMode.notifyDateTime && todoItem.endTime != "") {
-            val syncTime = appContext.getSp("todo").getLong("TODO_LAST_SYNC_TIME", 0L)
-            mViewModel.delTodo(DelPushWrapper(listOf(todoItem.todoId), syncTime, 1))
-            val currentList = todoAllAdapter.currentList.toMutableList()
-            currentList.remove(todoItem)
-            todoAllAdapter.submitList(currentList)
-        } else if (todoItem.endTime == "" && todoItem.remindMode.repeatMode != 0) {
-            var currentSystemTime = LocalDateTime.now()
-            if (todoItem.remindMode.notifyDateTime != "") {
-                currentSystemTime = todoItem.remindMode.notifyDateTime?.let { parseDateTime(it) }
-            }
-            val initialRemindTime = currentSystemTime
-            viewLifecycleOwner.lifecycleScope.launch {
-                val nextRemindTime = withContext(Dispatchers.Default) {
-                    when (todoItem.remindMode.repeatMode) {
-                        1 -> NocalculateNextDailyRemindTime(initialRemindTime)
-                        2 ->
-                            NocalculateNextWeeklyRemindTime(
-                                initialRemindTime,
-                                todoItem.remindMode.week
-                            )
-
-                        3 ->
-                            NocalculateNextMonthlyRemindTime(
-                                initialRemindTime,
-                                todoItem.remindMode.day
-                            )
-
-                        else -> initialRemindTime
-                    }
-                }
-
-                Log.d("current", "updateTodoItem: $nextRemindTime")
-                nextRemindTime?.let {
-                    if (todoItem.remindMode.notifyDateTime == "") {
-                        todoItem.remindMode.notifyDateTime = formatDateTime(it)
-                        Log.d("day", "updateTodoItem: $todoItem.remindMode.notifyDateTime")
-                        val syncTime = appContext.getSp("todo").getLong("TODO_LAST_SYNC_TIME", 0L)
-                        mViewModel.pushTodo(TodoListPushWrapper(listOf(todoItem), syncTime, 1, 0))
-                        todoAllAdapter.notifyDataSetChanged()
-                    } else {
-                        todoItem.remindMode.notifyDateTime = formatDateTime(it)
-                        val syncTime = appContext.getSp("todo").getLong("TODO_LAST_SYNC_TIME", 0L)
-                        mViewModel.pushTodo(TodoListPushWrapper(listOf(todoItem), syncTime, 1, 0))
-                        val currentList = todoAllAdapter.currentList.toMutableList()
-                        currentList.remove(todoItem)
-                        todoAllAdapter.submitList(currentList) {
-                            mRecyclerView.scrollToPosition(getTopItems())
-                        }
-                    }
-                }
-            }
-
-        } else {
-            var currentSystemTime = LocalDateTime.now()
-            if (todoItem.remindMode.notifyDateTime != "") {
-                currentSystemTime = todoItem.remindMode.notifyDateTime?.let { parseDateTime(it) }
-            }
-            val endTime = todoItem.endTime?.let { parseDateTime(it) }
-            val initialRemindTime = currentSystemTime
-
-            viewLifecycleOwner.lifecycleScope.launch {
-                val nextRemindTime = withContext(Dispatchers.Default) {
-                    when (todoItem.remindMode.repeatMode) {
-                        0 -> endTime
-                        1 -> endTime?.let { calculateNextDailyRemindTime(initialRemindTime, it) }
-                        2 -> endTime?.let {
-                            calculateNextWeeklyRemindTime(
-                                initialRemindTime,
-                                todoItem.remindMode.week,
-                                it
-                            )
-                        }
-
-                        3 -> endTime?.let {
-                            calculateNextMonthlyRemindTime(
-                                initialRemindTime,
-                                todoItem.remindMode.day,
-                                it
-                            )
-                        }
-
-                        else -> initialRemindTime
-                    }
-                }
-
-                Log.d("current", "updateTodoItem: $nextRemindTime")
-                nextRemindTime?.let {
-                    if (todoItem.remindMode.notifyDateTime == "") {
-                        todoItem.remindMode.notifyDateTime = formatDateTime(it)
-                        val syncTime = appContext.getSp("todo").getLong("TODO_LAST_SYNC_TIME", 0L)
-                        mViewModel.pushTodo(TodoListPushWrapper(listOf(todoItem), syncTime, 1, 0))
-                        todoAllAdapter.notifyDataSetChanged()
-                    } else {
-                        todoItem.remindMode.notifyDateTime = formatDateTime(it)
-                        val syncTime = appContext.getSp("todo").getLong("TODO_LAST_SYNC_TIME", 0L)
-                        mViewModel.pushTodo(TodoListPushWrapper(listOf(todoItem), syncTime, 1, 0))
-                        val currentList = todoAllAdapter.currentList.toMutableList()
-                        currentList.remove(todoItem)
-                        todoAllAdapter.submitList(currentList) {
-                            mRecyclerView.scrollToPosition(getTopItems())
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun calculateNextDailyRemindTime(
-        currentRemindTime: LocalDateTime,
-        endTime: LocalDateTime
-    ): LocalDateTime? {
-        var nextRemindTime =
-            currentRemindTime.plusDays(1).withHour(0).withMinute(0).withSecond(0).withNano(0)
-        if (nextRemindTime.isAfter(endTime)) {
-            return endTime
-        }
-        return nextRemindTime
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun NocalculateNextDailyRemindTime(
-        currentRemindTime: LocalDateTime
-    ): LocalDateTime? {
-        var nextRemindTime =
-            currentRemindTime.plusDays(1).withHour(0).withMinute(0).withSecond(0).withNano(0)
-        Log.d("day", "NocalculateNextDailyRemindTime:$nextRemindTime ")
-        return nextRemindTime
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun calculateNextWeeklyRemindTime(
-        currentRemindTime: LocalDateTime,
-        weekDays: List<Int>,
-        endTime: LocalDateTime
-    ): LocalDateTime? {
-        // 初始化为当前提醒时间并设置为当天的00:00:00
-        var nextRemindTime = currentRemindTime.withHour(0).withMinute(0).withSecond(0).withNano(0)
-
-        // 确保 weekDays 列表包含合法的值
-        val validWeekDays = weekDays.filter { it in 1..7 }
-        if (validWeekDays.isEmpty()) {
-            Log.w("ReminderCalculation", "No valid week days provided")
-            return null
-        }
-
-        // 找到下一个符合条件的周一
-        // 首先将 nextRemindTime 移动到第二天，确保不是当前日期
-        nextRemindTime = nextRemindTime.plusDays(1)
-
-        // 循环直到找到符合条件的下一个日期
-        while (nextRemindTime.dayOfWeek.value !in validWeekDays) {
-            nextRemindTime = nextRemindTime.plusDays(1)
-            if (nextRemindTime.isAfter(endTime)) {
-                Log.d("ReminderCalculation", "Next remind time is after end time")
-                return endTime // 超过截止时间，不再提醒
-            }
-        }
-
-        Log.d("运算结果", "calculateNextWeeklyRemindTime: $nextRemindTime")
-        return nextRemindTime
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun NocalculateNextWeeklyRemindTime(
-        currentRemindTime: LocalDateTime,
-        weekDays: List<Int>,
-    ): LocalDateTime? {
-        // 初始化为当前提醒时间并设置为当天的00:00:00
-        var nextRemindTime = currentRemindTime.withHour(0).withMinute(0).withSecond(0).withNano(0)
-
-        // 确保 weekDays 列表包含合法的值
-        val validWeekDays = weekDays.filter { it in 1..7 }
-        if (validWeekDays.isEmpty()) {
-            Log.w("ReminderCalculation", "No valid week days provided")
-            return null
-        }
-
-        // 找到下一个符合条件的周一
-        // 首先将 nextRemindTime 移动到第二天，确保不是当前日期
-        nextRemindTime = nextRemindTime.plusDays(1)
-
-        // 循环直到找到符合条件的下一个日期
-        while (nextRemindTime.dayOfWeek.value !in validWeekDays) {
-            nextRemindTime = nextRemindTime.plusDays(1)
-        }
-
-        Log.d("运算结果", "calculateNextWeeklyRemindTime: $nextRemindTime")
-        return nextRemindTime
-    }
-
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun calculateNextMonthlyRemindTime(
-        currentRemindTime: LocalDateTime,
-        days: List<Int>,
-        endTime: LocalDateTime
-    ): LocalDateTime? {
-        var nextRemindTime = currentRemindTime
-        val sortedDays = days.sorted() // 确保 days 是排序的
-
-        while (true) {
-            var found = false
-
-            // 遍历 days 中的每一天
-            for (day in sortedDays) {
-                val possibleRemindTime = nextRemindTime
-                    .withDayOfMonth(day)
-                    .withHour(0)
-                    .withMinute(0)
-                    .withSecond(0)
-                    .withNano(0)
-
-                if (possibleRemindTime.isAfter(currentRemindTime)) {
-                    nextRemindTime = possibleRemindTime
-                    found = true
-                    break // 找到下一个提醒时间，跳出循环
-                }
-            }
-
-            // 如果找到的时间在截止时间之前，则返回
-            if (found && nextRemindTime.isBefore(endTime)) {
-                Log.d("yue运算结果", "calculateNextMonthlyRemindTime: $nextRemindTime")
-                return nextRemindTime
-            }
-
-            // 如果没有找到合适的日期，增加一个月并重置为第一个日期
-            nextRemindTime = nextRemindTime.plusMonths(1).withDayOfMonth(sortedDays[0])
-
-            // 如果超过截止时间，返回截止时间
-            if (nextRemindTime.isAfter(endTime)) {
-                Log.d("yue运算结果", "calculateNextMonthlyRemindTime: $endTime")
-                return endTime
-            }
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun NocalculateNextMonthlyRemindTime(
-        currentRemindTime: LocalDateTime,
-        days: List<Int>,
-    ): LocalDateTime? {
-        var nextRemindTime = currentRemindTime
-        val sortedDays = days.sorted() // 确保 days 是排序的
-
-        // 循环直到找到下一个合适的提醒时间
-        while (true) {
-            var found = false
-
-            // 遍历 days 列表中的每一天
-            for (day in sortedDays) {
-                // 尝试将提醒时间设置为当前月的 day 号
-                val possibleRemindTime = nextRemindTime
-                    .withDayOfMonth(day)
-                    .withHour(0)
-                    .withMinute(0)
-                    .withSecond(0)
-                    .withNano(0)
-
-                // 如果找到的时间比当前时间晚，则返回该时间
-                if (possibleRemindTime.isAfter(currentRemindTime)) {
-                    nextRemindTime = possibleRemindTime
-                    found = true
-                    break // 找到合适的提醒时间，跳出循环
-                }
-            }
-
-            // 如果找到合适的提醒时间，跳出 while 循环
-            if (found) {
-                return nextRemindTime
-            }
-
-            // 如果没有找到，增加一个月并重置为下个月第一个指定的日期
-            nextRemindTime = nextRemindTime.plusMonths(1).withDayOfMonth(sortedDays[0])
-        }
-    }
-
-
-    private fun getTopItems(): Int {
-        var topitems = 0;
-        for (todo in todoAllAdapter.currentList) {
-            // 对每个 todo 对象执行操作
-            if (todo.isPinned == 1) {
-                topitems++
-            }
-        }
-        return topitems
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun parseDateTime(dateTimeStr: String): LocalDateTime {
-        return LocalDateTime.parse(dateTimeStr, dateTimeFormatter)
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun formatDateTime(dateTime: LocalDateTime): String {
-        return dateTime.format(dateTimeFormatter)
-    }
 
     @RequiresApi(Build.VERSION_CODES.O)
     override fun onFinishCheck(item: Todo) {
         pendingUpdateTask?.let { handler.removeCallbacks(it) }
-        if (item.remindMode.repeatMode != 0 ) {
+        if (item.remindMode.repeatMode != 0) {
             // 创建新的任务
             pendingUpdateTask = Runnable {
-                updateTodoItem(item)
+                updateTodoItem(
+                    item,
+                    requireContext(),
+                    mViewModel,
+                    todoAllAdapter,
+                    mRecyclerView,
+                    viewLifecycleOwner
+                )
             }
             // 延迟 1.5秒执行新的任务
             pendingUpdateTask?.let { handler.postDelayed(it, 1500) }
@@ -586,16 +272,36 @@ class TodoLifeFragment : BaseFragment(), TodoAllAdapter.OnItemClickListener {
 
     @RequiresApi(Build.VERSION_CODES.O)
     override fun onItemnotify(item: Todo) {
-        if (item.endTime != "") {
-            updateTodoItem(item)
+        if (item.endTime != "" && item.remindMode.repeatMode != 0) {
+            updateTodoItem(
+                item,
+                requireContext(),
+                mViewModel,
+                todoAllAdapter,
+                mRecyclerView,
+                viewLifecycleOwner
+            )
         } else {
             if (item.remindMode.repeatMode != 0) {
-                updateTodoItem(item)
+                updateTodoItem(
+                    item,
+                    requireContext(),
+                    mViewModel,
+                    todoAllAdapter,
+                    mRecyclerView,
+                    viewLifecycleOwner
+                )
+            } else {
+                item.remindMode.notifyDateTime = item.endTime
             }
+
         }
 
     }
 
+    fun getTopItems(): Int {
+        return todoAllAdapter.currentList.count { it.isPinned == 1 }
+    }
 
     override fun onResume() {
         mViewModel.getAllTodo()
